% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_function.R
\name{file_modification_time_cache}
\alias{file_modification_time_cache}
\title{Create function for operating on disk files with caching}
\usage{
file_modification_time_cache(fnc)
}
\arguments{
\item{fnc}{function that operates on file (first argument must be path to file).}
}
\value{
Return new function with caching functionality
}
\description{
Create function which cache result based on file. If file will be changed, function will also be recalculated, otherwise cached value will be returned
}
\examples{
# ADD EXAMPLES HERE

}

