% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawk.R
\name{rawk}
\alias{rawk}
\title{Create awk function.}
\usage{
rawk(code = "{print $1}")
}
\arguments{
\item{code}{awk code}
}
\value{
Retrun function which executes awk code on a file.

Returned function arguments:

\code{file} path to file

\code{stdout} standard output. By default result is printed into console. See \link{system2} for futher description.

\code{...} Other arguments passed to \link{system2}.
}
\description{
Create awk function.
}
\examples{

file = tempfile()
cat('1\\n2\\n', file = file)

rawk()(file)

# Capture output
awk = rawk('{print $0}')
awk(file, stdout = TRUE)

}

