% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systemAwk.R
\name{rawk_find_awk}
\alias{rawk_find_awk}
\title{Find awk installation.}
\usage{
rawk_find_awk()
}
\value{
Returns path to awk.

On Windows tries to use Sys.which("awk"), or Rtools gawk.

If it is unable to find rawk, returns NA.
}
\description{
This function is used for finding awk inplementation in the system.
}
\examples{

rawk_find_awk()

}

