% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systemAwk.R
\name{rawk_get_awk_path}
\alias{rawk_get_awk_path}
\title{Return path to awk}
\usage{
rawk_get_awk_path()
}
\value{
Returns path to awk. If it cannot find any awk, throws an error.
}
\description{
Return path to awk.
}
\examples{

rawk_get_awk_path()

}

