% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systemAwk.R
\name{rawk_set_awk_path}
\alias{rawk_set_awk_path}
\title{Set path to awk}
\usage{
rawk_set_awk_path(path, test = FALSE)
}
\arguments{
\item{path}{system path to awk implementations.}

\item{test}{if true, use simple test to determine if \code{path} is valid awk implementation.}
}
\value{
Returns invisible \code{path}.
}
\description{
Set path to awk for rawk calls.
}
\examples{

rawk_set_awk_path(rawk_get_awk_path(), test = TRUE)

}

