% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sed.R
\name{sed}
\alias{sed}
\title{Simple sed imitation.}
\usage{
sed(x, query, ...)
}
\arguments{
\item{x}{character vector}

\item{query}{sed query.}

\item{...}{other parameters passed to stringi functions.}
}
\value{
Character vector.
}
\description{
This function tries to imitate sed (stream editor), especially \code{s/foo/bar/g} syntax.
}
\examples{

query = "s/foo/bar/g"
x = "foobar"
sed(x, query)

}

